ДЗ№7. Департамент ATM. Правки
