+++ hw07-ATMDepartment/src/main/java/ru/otus/hw07/ATMDepartment.java
@@ -1,28 +1,20 @@
 package ru.otus.hw07;
 
-import ru.otus.hw07.ATM.ATM;
-import ru.otus.hw07.visitor.ATMVisitor;
+/**
+ * Интерфейс департамента АТМ.
+ */
 
-import java.util.ArrayList;
+public interface ATMDepartment {
+    /**
+     * Получение суммы остатков во всех банкоматах
+     *
+     * @return сумма остатков во всем банкоматах
+     */
+    Integer getATMAmounts();
 
-public class ATMDepartment implements ATMDepartmentInterface {
-    private final ArrayList<ATM> atms;
+    /**
+     * Возврат состояния всех банкоматов к начальному состоянию.
+     */
+    void restoreInitialStates();
 
-    public ATMDepartment(ArrayList<ATM> atms) {
-        this.atms = atms;
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public Integer getATMAmounts() {
-        return this.atms.stream().map(atm -> atm.accept(new ATMVisitor())).reduce(Integer::sum).orElse(0);
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public void restoreInitialStates() {
-        for(ATM atm: this.atms){
-            atm.restoreDefaultATM();
-        }
-    }
 }
