+++ hw07-ATMDepartment/src/main/java/ru/otus/hw07/visitor/ATMVisitorImpl.java
@@ -1,0 +1,13 @@
+package ru.otus.hw07.visitor;
+
+import ru.otus.hw07.ATM.ATMImpl;
+
+public class ATMVisitorImpl implements ATMVisitor {
+
+    /** {@inheritDoc} */
+    @Override
+    public Integer visit(ATMImpl atm) {
+       return atm.getTotalAmountInNomunal();
+    }
+
+}
